moved GetNextStatusTest.php and GetAvailableActionTest.php into one file TaskTest and add testing of exceptions also
